requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cryptography==46.0.5
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.128.8
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.4.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
